images: dash as scripting shell, shell commands
When replacing busybox with toybox (to be committed separately),
several commands will disappear from the base image (ping, top, ps,
vi, etc.).

This is acceptable for a base image, but not for something that people
use interactively or the QA test scripts. Therefore a new
"tools-interactive" image feature gets introduced:
- it is active in all image variants at the moment
- the corresponding package group installs commands not provided
  anymore by toybox, choosing the corresponding individual
  packages instead
- bash is installed but not used as /bin/sh, which always points
  to dash

The goal is to ensure consistent behavior, regardless whether
coreutils, busybox or toybox are installed.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>