images: dash as scripting shell, shell commands
When replacing busybox with toybox (to be committed separately),
several commands will disappear from the base image (ping, top, ps,
vi, etc.).

This is acceptable for a base image, but not for something that people
use interactively or the QA test scripts. Therefore a new
"tools-interactive" image feature gets introduced:
- it is active in all image variants at the moment
- the corresponding package group installs commands not provided
  anymore by toybox, choosing the corresponding individual
  packages instead
- bash is installed and set as login shell for the root account
  and in the template used for creating new users

The goal is to ensure consistent behavior, regardless whether
coreutils, busybox or toybox are installed. Therefore dash is
installed as /bin/sh unconditionally. In future images that
provide no shell it can be removed again.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>