ostro.conf: ignore initramfs-framework->toybox dependency
The sstate selftest warns about the dependency unless it gets
explicitly marked as ABI-safe. OE-core does the same for
initramfs-framework->busybox, which makes sense because the "ABI" are
the shell commands.

A fix for this will also be sent to OE-core, in which case this one
here can be reverted.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>