stateless: initial support class and configuration
The approach taken here centralizes the necessary package and image
modifications in a single support class and a single configuration
file.

The alternative would be to introduce a "stateless" distro feature and
then modifify the various upstream recipes and helper classes to
support it. This would be doable, but could be tricky to get accepted
into all the various upstream components. It probably also would
require more changes. With the current approach, only some minor fixes
are needed in the original recipes.

The current configuration is good enough to build and run
"ostro-image-minimal" under qemu, with Smack and IMA enabled. However,
/etc is not completely empty yet: several exceptions were whitelisted
and marked as TODO in stateless.inc. These can now be addresssed one
by one collaboratively.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>