ca-certificates: work around Toybox segfault and sed limitations
Toybox 0.7.0 sed segfaults when invoked with "-e '$a\'"; the patch
mentions a certain .crt, but it also happens when applied to
/dev/null. A commit in Toybox master solves the segfault, but the
Toybox (incorrectly?!) complains about the lack of text after "a\".

We work around it by using a simpler combination of "cat + echo". That
always adds another newline in contrast to the solution using sed, but
that shouldn't matter that much.

Signed-off-by: Patrick Ohly <patrick.ohly@intel.com>