Move wpa_supplicant files from etc to defaults dir
Let's just move wpa_supplicant.conf to defaults dir.
Currently the file is not used and it is the responsibility
of the one who starts the wpa_supplicant daemon to give the
conf file as parameter to daemon. Currently we are starting
wpa_supplicant with dbus interface, which Connman is then
using. At least in Galileo the moving didn't make any
difference to current wlan operation.

Move also the the wpa-supplicant dbus conf file to
/usr/share/dbus-1 because our version of d-bus should
be able to read from there also.

Remove for now the ifup and ifdown related wpa_supplicant
scripts because we are not using them and they contain
hardcoded paths to /etc/interfaces.

Signed-off-by: Jaska Uimonen <jaska.uimonen@intel.com>